pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robotsix-agents"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.0",
    "platformdirs>=4.4.0",
    "pyyaml>=6.0",
    "types-pyyaml>=6.0",
    "autogen-core>=0.7.4",
    "autogen-agentchat>=0.7.4",
    "autogen-ext[mem0]",
    "tiktoken",
    "ollama",
    "caldav",
    "mcp-server-git",
    "cocoindex[embeddings]",
    "docker",
    "psycopg2-binary",
]

[project.scripts]
ra-config = "robotsix_agents.core.config_manager:main"
ra-orchestrator = "robotsix_agents.orchestrator.cli:main"

[tool.setuptools.packages.find]
where = ["."]
